requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
certifi==2024.8.30
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
distlib==0.3.9
distro==1.9.0
docstring_parser==0.16
filelock==3.16.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
identify==2.6.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
instructor==1.6.3
Jinja2==3.1.4
jiter==0.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
moviepy==1.0.3
multidict==6.1.0
nodeenv==1.9.1
numpy==2.1.2
openai==1.53.0
packaging==24.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.0.1
proglog==0.1.10
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.3
setuptools==75.3.0
shellingham==1.5.4
sniffio==1.3.1
swarm @ git+https://github.com/openai/swarm.git@9db581cecaacea0d46a933d6453c312b034dbf47
tenacity==9.0.0
tqdm==4.66.6
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
virtualenv==20.27.1
yarl==1.17.1
